+++ RELEASE_NOTES.md
@@ -12,7 +12,7 @@
 
   `Receiver.consume()` doesn't raise any exceptions.
 
-  Receivers raising `EOFError` now raise `ReceiverError` instead.
+  Receivers raising `EOFError` now raise `ReceiverInvalidatedError` instead.
 
 * For channels which senders raise an error when the channel is closed or which receivers stop receiving when the channel is closed, the `SenderError` and `ReceiverStoppedError` are chained with a `__cause__` that is a `ChannelClosedError` with the channel that was closed.
 
@@ -29,6 +29,8 @@
   * `ReceiverError`: Raised for errors when receiving messages.
 
   * `ReceiverClosedError`: Raised when a receiver don't have more messages to receive.
+
+  * `ReceiverInvalidatedError`: Raised when a receiver was invalidated (for example it was converted into a `Peekable`).
 
 ## Bug Fixes